chore(vs-component): moved RTLComponent, DRRAError and ParameterList to a new types.rs in models
