+++ modules/vs-component/src/models.rs
@@ -4,3 +4,5 @@ pub mod controller;
 pub mod drra;
 pub mod pcu;
 pub mod resource;
+
+pub mod types;
